requirements.txt
anyio==3.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
fastapi==0.95.1
filelock==3.12.0
h11==0.14.0
idna==3.4
Jinja2==3.1.2
llvmlite==0.40.0
MarkupSafe==2.1.2
more-itertools==9.1.0
mpmath==1.3.0
networkx==3.1
numba==0.57.0
numpy==1.24.3
openai-whisper @ git+https://github.com/openai/whisper.git@248b6cb124225dd263bb9bd32d060b6517e067f8
PyAudio==0.2.13
pydantic==1.10.7
regex==2023.5.5
requests==2.30.0
sniffio==1.3.0
SpeechRecognition==3.10.0
starlette==0.26.1
sympy==1.11.1
tiktoken==0.3.3
torch==2.0.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==2.0.2
uvicorn==0.22.0
